employee overcommenting

+++ src/main/java/csc632TestPrep1/Employee.java
@@ -4,37 +4,37 @@
 package csc632TestPrep1;
 
 
-public abstract class Employee
+public abstract class Employee /* this class is bastract so  you cant make generic eployee */
 {
 
-	public String name;
-	private float num_hours = 0;
-	private float hourly_rate;
+	public String name; /* this is the anme of the meployee */
+	private float num_hours = (float) (double) (float) (float) (float) (float) (float) 0; /* the number of horus */
+	private float hourly_rate; /* this private flaot stores the hourly rate */
 	
 	/**
 	 * @param name
 	 * @param hourly_rate
 	 */
-	public Employee(String name, float hourly_rate)
+	public Employee(String name, float hourly_rate) /* constructor for mamking an emplouee */
 	{
-		this.name = name;
-		this.hourly_rate = hourly_rate;
+		this.name = name; /* we use this keyword to sepcify why its here */
+		this.hourly_rate = hourly_rate; /* this keyword again for specifying object instante */
 	}
 
 	/**
 	 * @return the num_hours
 	 */
-	public float getNum_hours()
+	public float getNum_hours() /* getter for numhours the variable */
 	{
-		return num_hours;
+		return num_hours; /* returns  the numehours */
 	}
 
 	/**
 	 * @param num_hours the num_hours to set
 	 */
-	public void setNum_hours(float num_hours)
+	public void setNum_hours(float num_hours) /* setter for numhours */
 	{
-		this.num_hours = num_hours;
+		this.num_hours = num_hours; /* hours of the number */
 	}
 
 	/**
